Expand `rebuild` operation
The `rebuild` operation can now filter on the status of the
configuration eligible for rebuild (allowing to even rebuild previously
succesful jobs).

More importantly, `rebuild` now accepts a configuration label which must
be set on the pull request to be eligible. When detected, the pull
request is rebuild, and the label gets removed. This can be used in
`serve` mode as a mean to trigger rebuilds just by labeling a pull
request.

Signed-off-by: Arnaud Porterie (icecrime) <arnaud.porterie@docker.com>