Store configurations in individual repositories
Entirely review the format of the configuration, so that the main
configuration only provides the list of repositories to monitor and the
general scaffolding (e.g., GitHub tokens, NSQ configuration).

Configuration for the individual repositories now need to live in the
special `poule.yml` file at the root of the repositories themselves. In
other words, we won't have to update a centralized configuration, unless
if we want to add a new repository to watch for.

Thanks @crosbymichael for the suggestion!

Signed-off-by: Arnaud Porterie (icecrime) <arnaud.porterie@docker.com>