Ignore failure to remove label in `dco-check`
We're getting a lot of errors in the log as `dco-check` always attempts
to remove the `dco/no` label, even when absent. This produces a 404 in
the GitHub API, and is treated by the operation as an error (hence, the
message gets requeued in NSQ).

Safely ignore 404 when trying to remove the `dco/no` label.

Signed-off-by: Arnaud Porterie (icecrime) <arnaud.porterie@docker.com>